Convert description & comment body to Markdown. Fixes #15
This is a crazy and roundabout way of doing things but was the only
way I could find to do it.

First we use an undocumented Jira API call to render the Jira markup
as HTML. Then we use the reverse_markdown gem to convert the HTML to
Markdown. Phew.